Remove @UiThread and @WorkerThread from public listeners
Setting up a listener in ui thread incorrectly triggers a warning
of not being in a worker thread, it's just an interface but code
analysis makes wrong assumptions. It also cause the same bug vice versa
if you setup a listener from worker thread and get warnings for ui thread.
Setting up a listener is not the same as it being invoked, setting up
is allowed in any thread but code analysis not smart enough.

Plus people that don't have google annotations as a dependency can get
compiler warnings anyway so it was already established not to use
the annotations in public APIs.